Bug fix; code refactor; grid analysis feature added
Fixed a bug with the adjacency functions; fixed a bug in the
command-line environment when executing the program on the GPU.

Additionally, refactored the grid to have "dead space" in last row and
column as opposed to first row and column. This removes the necessity of
constantly adding 1 to indices of the grid.

Grid analysis methods have been added; these count the number of live
cells, and return the fraction of neighbors of live cells that are also
live.